#11-9 CBV_SRV_UAVのディスクリプタヒープクラス追加 テクスチャマッピング実装完了

+++ RenderingEngine/src/DescriptorHeapRTV.h
@@ -20,7 +20,7 @@ public:
 
 private:
 	Microsoft::WRL::ComPtr<ID3D12DescriptorHeap> _rtvHeap = nullptr;	// ディスクリプタヒープ
-	UINT _handleIncrimentSize = 0;										// ハンドルのインクリメントサイズ
+	SIZE_T _handleIncrimentSize = 0;										// ハンドルのインクリメントサイズ
 	UINT _registedRTVNum = 0;									// ヒープに登録されたディスクリプタ数
 	UINT _nextHandleLocation = 1;
 
@@ -61,7 +61,7 @@ public:
 	/// ディスクリプタヒープハンドルのインクリメントサイズ取得
 	/// </summary>
 	/// <returns></returns>
-	UINT GetHandleIncrimentSize() {
+	SIZE_T GetHandleIncrimentSize() {
 		return _handleIncrimentSize;
 	}
